+++ src/netplay/netplay_session.rs
@@ -130,7 +130,7 @@ impl NetplaySession {
             }
         }
 
-        *Emulator::emulation_speed().write().unwrap() = if sess.frames_ahead() > 0 {
+        *Emulator::emulation_speed_mut() = if sess.frames_ahead() > 0 {
             log::debug!(
                 "Frames ahead: {:?}, slowing down emulation",
                 sess.frames_ahead()
