+++ src/window/winit_impl/mod.rs
@@ -1,6 +1,7 @@
-use winit::dpi::LogicalSize;
-
-use crate::input::keys::{KeyCode, Modifiers};
+use crate::{
+    input::keys::{KeyCode, Modifiers},
+    integer_scaling::MINIMUM_INTEGER_SCALING_SIZE,
+};
 
 use super::Fullscreen;
 
@@ -19,10 +20,7 @@ impl Fullscreen for winit::window::Window {
             use winit::platform::macos::WindowExtMacOS;
             if window.simple_fullscreen() {
                 window.set_simple_fullscreen(false);
-                let _ = window.request_inner_size(LogicalSize::new(
-                    crate::MINIMUM_INTEGER_SCALING_SIZE.0,
-                    crate::MINIMUM_INTEGER_SCALING_SIZE.1,
-                ));
+                let _ = window.request_inner_size(MINIMUM_INTEGER_SCALING_SIZE);
             } else {
                 window.set_simple_fullscreen(true);
             }
@@ -35,10 +33,7 @@ impl Fullscreen for winit::window::Window {
         {
             if window.fullscreen().is_some() {
                 window.set_fullscreen(None);
-                let _ = window.request_inner_size(LogicalSize::new(
-                    crate::MINIMUM_INTEGER_SCALING_SIZE.0,
-                    crate::MINIMUM_INTEGER_SCALING_SIZE.1,
-                ));
+                let _ = window.request_inner_size(MINIMUM_INTEGER_SCALING_SIZE);
             } else {
                 window.set_fullscreen(Some(winit::window::Fullscreen::Borderless(None)));
             }
